Add skeleton for an asdl parser
This would be a first potentially-useful project; in principle
the Pyre team would find a Rust generator for asdl grammars handy.

We don't really need the generator itself to be written in Rust,
but given that I'm trying to do some Rust tinkering in my spare
time I might as well give it a shot; compared to any full-fledged
programming language the grammar at least is relatively simple.